pyproject.toml
[project]
name = "imweb-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.5",
    "fastapi>=0.116.1",
    "uvicorn>=0.24.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.26.0",
    "pyjwt>=2.8.0",
    "requests>=2.31.0",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0"
]
